teach cppunit about known failures of tests
This way a tester can implement expected behavior that is known to fail.
For example in a test driven situation when failing tests are first
implemented they can be marked as such so they do not look like regressions.

Or a tester can implement tests for found bugs to document them first and
the fixes can come later when the priority is high enough.